refactor: layout을 adapter가 아닌 viewHolder에서 그리도록 수정
