+++ app/src/main/java/woowacourse/movie/ui/home/adapter/MovieListAdapter.kt
@@ -1,11 +1,7 @@
 package woowacourse.movie.ui.home.adapter
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.movie.R
-import woowacourse.movie.databinding.ItemAdBinding
-import woowacourse.movie.databinding.ItemMovieBinding
 import woowacourse.movie.uimodel.MovieListModel
 
 class MovieListAdapter(
@@ -16,16 +12,12 @@ class MovieListAdapter(
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         return when (viewType) {
             MovieListViewType.MOVIE.value -> {
-                val view =
-                    LayoutInflater.from(parent.context).inflate(R.layout.item_movie, parent, false)
-                MovieItemViewHolder(ItemMovieBinding.bind(view)) {
+                MovieItemViewHolder(parent) {
                     onItemClick.onMovieItemClick(modelItems[it] as MovieListModel.MovieModel)
                 }
             }
             MovieListViewType.AD.value -> {
-                val view =
-                    LayoutInflater.from(parent.context).inflate(R.layout.item_ad, parent, false)
-                AdItemViewHolder(ItemAdBinding.bind(view)) {
+                AdItemViewHolder(parent) {
                     onItemClick.onAdItemClick(modelItems[it] as MovieListModel.AdModel)
                 }
             }
